Cleaned up class structure
