+++ src/main/java/cambio/simulator/export/MultiDataPointReporter.java
@@ -1,8 +1,6 @@
 package cambio.simulator.export;
 
 import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.*;
 
 import desmoj.core.simulator.Model;
 import desmoj.core.simulator.TimeInstant;
@@ -13,41 +11,24 @@ import org.jetbrains.annotations.NotNull;
  *
  * @author Lion Wagner
  */
-public class MultiDataPointReporter extends MiSimReporter {
-
-    protected final String datasetsPrefix;
-    protected final Path reportBasePath;
-    protected final HashMap<String, AsyncReportWriter<?>> writerThreads = new HashMap<>();
-
-    protected final HashMap<String, String[]> customHeaders = new HashMap<>();
+public class MultiDataPointReporter extends MiSimReporter<AsyncMultiColumnReportWriter> {
 
     public MultiDataPointReporter(@NotNull Model model) {
         this("", model);
     }
 
     public MultiDataPointReporter(@NotNull String datasetsPrefix, @NotNull Model model) {
-        super(model);
-        Objects.requireNonNull(datasetsPrefix);
-
-        this.datasetsPrefix = datasetsPrefix;
-        this.reportBasePath = this.model.getExperimentMetaData().getReportLocation().resolve("raw");
+        super(model, datasetsPrefix);
     }
 
-    /**
-     * Adds a new datapoint to the given dataset.
-     *
-     * @param dataSetName name of the dataset to which the datapoint should be added
-     * @param when        point in simulation time to which the datapoint is associated to
-     * @param data        data that should be logged
-     * @param <T>         type of the data that should be logged.
-     */
+    @Override
     public <T> void addDatapoint(final String dataSetName, final TimeInstant when, final T... data) {
         checkArgumentsAreNotNull(dataSetName, when, data);
 
         if (!writerThreads.containsKey(dataSetName)) {
             try {
                 Files.createDirectories(reportBasePath);
-                AsyncReportWriter<?> writerThread =
+                AsyncMultiColumnReportWriter writerThread =
                     new AsyncMultiColumnReportWriter(
                         reportBasePath.resolve(datasetsPrefix + dataSetName + ".csv"),
                         customHeaders.getOrDefault(dataSetName,
@@ -58,31 +39,7 @@ public class MultiDataPointReporter extends MiSimReporter {
             }
         }
 
-        AsyncReportWriter<?> writerThread = writerThreads.get(dataSetName);
+        AsyncMultiColumnReportWriter writerThread = writerThreads.get(dataSetName);
         writerThread.addDataPoint(when.getTimeAsDouble(), data);
-    }
-
-
-    @Override
-    public void finalizeReport() {
-        writerThreads.values().forEach(AsyncReportWriter::finalizeWriteout);
-        super.deregister();
-        customHeaders.clear();
-    }
-
-    @Override
-    public String toString() {
-        return this.getClass().getSimpleName()
-            + (datasetsPrefix.equals("") ? "" : "{"
-            + "datasetsPrefix='" + datasetsPrefix + '\''
-            + '}');
-    }
-
-    public void registerDefaultHeader(String dataSetName, String... headers) {
-        if (customHeaders.putIfAbsent(dataSetName, headers) == null) {
-            throw new IllegalArgumentException(
-                "Header for dataset " + dataSetName + " already registered as "
-                    + Arrays.toString(customHeaders.get(dataSetName)));
-        }
     }
 }
